requirements.txt
alabaster==0.7.10
alembic==0.9.1
aniso8601==1.2.0
appdirs==1.4.3
Babel==2.3.4
bcrypt==3.1.3
cffi==1.9.1
click==6.7
docutils==0.13.1
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
imagesize==0.7.1
itsdangerous==0.24
Jinja2==2.9.5
Mako==1.0.6
MarkupSafe==1.0
packaging==16.8
pycparser==2.17
pyftpdlib==1.5.2
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.2.0
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2016.10
requests==2.13.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.3
SQLAlchemy==1.1.6
Werkzeug==0.12.1
